// lib/types.ts
interface UserState {
  username: string | null;
  role: string | null;
  companyName: string | null;
  vendorType: string | null;
  isLoggedIn: boolean;
  login: (userData: {
    username: string;
    role: string;
    companyName: string | null;
    vendorType: string | null;
  }) => void;
  logout: () => void;
}

export const useAuthStore = create<UserState>()(
  persist(
    (set) => ({
      username: null,
      role: null,
      companyName: null,
      vendorType: null,
      isLoggedIn: false,

      login: (userData) =>
        set({
          username: userData.username,
          role: userData.role,
          companyName: userData.companyName,
          vendorType: userData.vendorType,
          isLoggedIn: true,
        }),

      logout: () =>
        set({
          username: null,
          role: null,
          companyName: null,
          vendorType: null,
          isLoggedIn: false,
        }),
    }),
    {
      name: 'auth-storage',
      storage: createJSONStorage(() => localStorage),
    }
  )
);
// ...
export interface BomMaterialItem {
  id: number;
  material: string;
  materialDescription: string;
  partReference: string;
  qty: number;
}